+++ datalog/src/dev/java/edu/wpi/first/datalog/DevMain.java
@@ -1,0 +1,12 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package edu.wpi.first.datalog;
+
+public final class DevMain {
+  /** Main entry point. */
+  public static void main(String[] args) {}
+
+  private DevMain() {}
+}
